Move test utilities to new module `olmo_core.testing`
