+++ src/olmo_core/testing/__init__.py
@@ -1,0 +1,43 @@
+from .distributed import run_distributed_test
+from .utils import (
+    BACKENDS,
+    DEVICES,
+    FLASH_MARKS,
+    GPU_MARKS,
+    GROUPED_GEMM_MARKS,
+    INIT_DEVICES,
+    LOW_PRECISION_DTYPES,
+    MULTI_GPU_MARKS,
+    get_default_device,
+    has_cuda,
+    has_flash_attn,
+    has_grouped_gemm,
+    has_multiple_gpus,
+    has_torchao,
+    requires_flash_attn,
+    requires_gpu,
+    requires_grouped_gemm,
+    requires_multi_gpu,
+)
+
+__all__ = [
+    "BACKENDS",
+    "DEVICES",
+    "FLASH_MARKS",
+    "GPU_MARKS",
+    "GROUPED_GEMM_MARKS",
+    "INIT_DEVICES",
+    "LOW_PRECISION_DTYPES",
+    "MULTI_GPU_MARKS",
+    "get_default_device",
+    "has_cuda",
+    "has_flash_attn",
+    "has_grouped_gemm",
+    "has_multiple_gpus",
+    "has_torchao",
+    "requires_flash_attn",
+    "requires_gpu",
+    "requires_grouped_gemm",
+    "requires_multi_gpu",
+    "run_distributed_test",
+]
